requirements.txt
weasyprint
jinja2
pydantic
google-generativeai
fastapi
uvicorn
matplotlib
python-dotenv
